+++ src/action/input-check/index.ts
@@ -5,16 +5,25 @@ export const tempService = async (formdata: FormData) => {
 };
 
 export const paymentTempService = async (formdata: FormData) => {
-  const addressUuid = formdata.get('addressUuid');
-  const cardUuid = formdata.get('paymentCardUuid');
-  const product = [
-    {
-      productCode: formdata.get(`orderList[${0}].productCode`),
-      productName: formdata.get(`orderList[${0}].productName`),
-      quantity: formdata.get(`orderList[${0}].quantity`),
-      productPrice: formdata.get(`orderList[${0}].productPrice`),
-    },
-  ];
+  // const addressUuid = formdata.get('addressUuid');
+  // const cardUuid = formdata.get('paymentCardUuid');
+  // const productCode = formdata.get('orderList[0].productCode');
+  // const productName = formdata.get('orderList[0].productName');
+  // const quantity = formdata.get('orderList[0].quantity');
+  // const productPrice = formdata.get('orderList[0].productPrice');
 
-  console.log('결제 정보 테스트', addressUuid, cardUuid, product);
+  // console.log(
+  //   '결제 정보 테스트',
+  //   addressUuid,
+  //   cardUuid,
+  //   productCode,
+  //   productName,
+  //   quantity,
+  //   productPrice
+  // );
+  console.log('🧾 전체 결제 정보');
+
+  for (const [key, value] of formdata.entries()) {
+    console.log(`${key}: ${value}`);
+  }
 };
