+++ src/types/PaymentDataType.ts
@@ -42,6 +42,7 @@ export interface StarbuckscardChargeOptionType {
 
 export interface CartOrderItemInfoType {
   cartUuid: string;
+
   productCode: string;
   productOptionListId: number;
   quantity: number;
@@ -78,6 +79,7 @@ export interface ShowOrderProductDataType {
   productOptionListId: number[];
   quantity: number;
   thumbnailUrl: string;
+  productTotalPrice: number;
 }
 
 export interface ProductOptionDataType {
